backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
authlib==1.2.1
httpx==0.25.2

# Data Processing & ML
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# Sentiment Analysis
transformers==4.35.2
torch==2.1.1
vaderSentiment==3.3.2

# External APIs
praw==7.7.1
yfinance==0.2.33
requests==2.31.0
aiohttp==3.9.1

# Task Queue
celery==5.3.4
flower==2.0.1

# WebSocket
python-socketio==5.10.0

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0

# Monitoring
prometheus-client==0.19.0
